respect selected view when clicking on new task (automagik-forge 10f15d45)
When we click on a task in kanban, it is opnening the task in full view mode.

However it should respect thevieew;

eg:

User has a task open with preview mode, when it clicks in another task in kanban it should open the task with preview view opened.

WHen user has the kanban open, should open the task with the  kanban view as wll