Quick fixes QoL (#200)
* remove preview button in summary & actions (automagik-forge 1d82bed0)

currently we ahve a preview button in summary n actions
 ![{C681CEC2-0581-49A7-9892-8B76D7BE54C0}.png](.forge-images/11c31d1a-1344-491d-bd0e-12adcabea35e.png)

___

what we can do instead here is, rmeove htis button and add as behavior that. when we start the server it automatically changes the UI to preview mode

* On onboarding fix provider dropdown (automagik-forge 18ac8398)

![{1AED48C3-B963-4CE5-8DE3-BC694A915CEB}.png](/api/images/24478df8-8c8c-486f-a582-70009d8fc5e2/file)

This dropdown was recently changed, we should have unified (Settings, Chat, New Task) however in this point where we are in onboarding, it is still showing old stuff.
Exampel of new one:

 ![{F0A21085-E2BF-4606-8FDB-D2DAB61D7A34}.png](/api/images/44c8cb18-d4b1-4061-a009-d2fa9d6d9056/file)

* respect selected view when clicking on new task (automagik-forge 10f15d45)

When we click on a task in kanban, it is opnening the task in full view mode.

However it should respect thevieew;

eg:

User has a task open with preview mode, when it clicks in another task in kanban it should open the task with preview view opened.

WHen user has the kanban open, should open the task with the  kanban view as wll

* On onboarding fix provider dropdown (automagik-forge 18ac8398)

![{1AED48C3-B963-4CE5-8DE3-BC694A915CEB}.png](/api/images/24478df8-8c8c-486f-a582-70009d8fc5e2/file)

This dropdown was recently changed, we should have unified (Settings, Chat, New Task) however in this point where we are in onboarding, it is still showing old stuff.
Exampel of new one:

 ![{F0A21085-E2BF-4606-8FDB-D2DAB61D7A34}.png](/api/images/44c8cb18-d4b1-4061-a009-d2fa9d6d9056/file)

* fix: approval with rebase (automagik-forge 2797109d)

When a button with "Approval" shows and there is something to rebase, there's no indication that the user needs to click on rebase to continue the approval.

We need to addres this issue

* respect selected view when clicking on new task (automagik-forge 10f15d45)

When we click on a task in kanban, it is opnening the task in full view mode.

However it should respect thevieew;

eg:

User has a task open with preview mode, when it clicks in another task in kanban it should open the task with preview view opened.

WHen user has the kanban open, should open the task with the  kanban view as wll

* fix: task finish notification sound (automagik-forge c0a1faca)

Curerently there's no way to set the volume of the notification, and its too fucking loud.

We should add a slider/option to set the default value of the sound when a task finishes.

Also we should set the sound to be 30% by default

* respect selected view when clicking on new task (automagik-forge 10f15d45)

When we click on a task in kanban, it is opnening the task in full view mode.

However it should respect thevieew;

eg:

User has a task open with preview mode, when it clicks in another task in kanban it should open the task with preview view opened.

WHen user has the kanban open, should open the task with the  kanban view as wll

* fix: in breadcrumb board should have the icon (automagik-forge 7cc36843)

Currently the board icon is in the project name and not in the Actual "Board" part, this is really misleading.

* fix: task finish notification sound (automagik-forge c0a1faca)

Curerently there's no way to set the volume of the notification, and its too fucking loud.

We should add a slider/option to set the default value of the sound when a task finishes.

Also we should set the sound to be 30% by default

* fix: approval with rebase (automagik-forge 2797109d)

When a button with "Approval" shows and there is something to rebase, there's no indication that the user needs to click on rebase to continue the approval.

We need to addres this issue

* respect selected view when clicking on new task (automagik-forge 10f15d45)

When we click on a task in kanban, it is opnening the task in full view mode.

However it should respect thevieew;

eg:

User has a task open with preview mode, when it clicks in another task in kanban it should open the task with preview view opened.

WHen user has the kanban open, should open the task with the  kanban view as wll

* Chat Copy Button (automagik-forge 80e7e55e)

Copy chat in a specific .md format

USer messages
Assistant messages
Assistant tool calls
Etc whatever is possible, Simple button where would be the best place?  maybe inside the history add a download button or smth not sure

_____

Let's make a structured thing. kinda like frontmatter, where at the top we have some metadata info , ids etc. and a proper fromatt where depending on how we do it , we could "import" this chat

* feat: highlight selected taskcard in kanban (automagik-forge 73377735)

When in task open like http://127.0.0.1:9000/projects/37ec2870-2dd1-4f97-ba1b-2bd2bc46b676/tasks/34f76600-5369-4129-84cb-88d4ac7cb89d/attempts/d1be8c8f-bc91-4c3e-91f4-a23b334bb804?view=kanban
And we set the view to kanban, highlight the task in the board